separate out table header and table, refactor to pass less into child components, clean up styles more
